Refactor to transaction file
